test: consolidate safety tests and add pointer-equality and backup/restore coverage
- Merge all separate *_safety_test.go files into existing test files to
  reduce fragmentation (rdner review feedback)
- Add assert.Same checks to verify no event.Clone() occurs when no copy
  is needed
- Add TestReplaceMultiFieldBackup and TestTruncateFieldsMultiFieldBackup
  to verify multi-field backup/restore correctly undoes partial changes
  on error
- Add missing error message assertions in append safety tests
