+++ libbeat/processors/actions/copy_fields_test.go
@@ -20,12 +20,13 @@ package actions
 import (
 	"testing"
 
-	"github.com/elastic/elastic-agent-libs/logp/logptest"
-	"github.com/elastic/elastic-agent-libs/mapstr"
-
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 
 	"github.com/elastic/beats/v7/libbeat/beat"
+	conf "github.com/elastic/elastic-agent-libs/config"
+	"github.com/elastic/elastic-agent-libs/logp/logptest"
+	"github.com/elastic/elastic-agent-libs/mapstr"
 )
 
 func TestCopyFields(t *testing.T) {
@@ -199,4 +200,73 @@ func TestCopyFields(t *testing.T) {
 		assert.Equal(t, expMeta, newEvent.Meta)
 		assert.Equal(t, event.Fields, newEvent.Fields)
 	})
+}
+
+// TestCopyFieldsSingleFieldNoClone verifies that a single-field copy
+// produces identical results without cloning.
+func TestCopyFieldsSingleFieldNoClone(t *testing.T) {
+	tests := []struct {
+		name   string
+		from   string
+		to     string
+		input  mapstr.M
+		output mapstr.M
+		err    bool
+	}{
+		{
+			name:  "happy path: message to event.original",
+			from:  "message",
+			to:    "event.original",
+			input: mapstr.M{"message": "hello world"},
+			output: mapstr.M{
+				"message": "hello world",
+				"event":   mapstr.M{"original": "hello world"},
+			},
+		},
+		{
+			name:   "source missing",
+			from:   "nonexistent",
+			to:     "dest",
+			input:  mapstr.M{"message": "hello"},
+			output: mapstr.M{"message": "hello"},
+			err:    true,
+		},
+		{
+			name:   "target already exists",
+			from:   "message",
+			to:     "existing",
+			input:  mapstr.M{"message": "hello", "existing": "already here"},
+			output: mapstr.M{"message": "hello", "existing": "already here"},
+			err:    true,
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			c, err := conf.NewConfigFrom(map[string]interface{}{
+				"fields": []map[string]interface{}{
+					{"from": tc.from, "to": tc.to},
+				},
+				"fail_on_error": true,
+			})
+			require.NoError(t, err)
+
+			p, err := NewCopyFields(c, logptest.NewTestingLogger(t, ""))
+			require.NoError(t, err)
+
+			event := &beat.Event{Fields: tc.input.Clone()}
+			result, err := p.Run(event)
+			assert.Same(t, event, result)
+
+			if tc.err {
+				assert.Error(t, err)
+			} else {
+				assert.NoError(t, err)
+			}
+
+			result.Fields.Delete("error")
+			assert.Equal(t, tc.output, result.Fields,
+				"single-field copy must produce same result without clone")
+		})
+	}
 }
