+++ libbeat/processors/extract_array/extract_array_test.go
@@ -21,6 +21,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 
 	"github.com/elastic/beats/v7/libbeat/beat"
 	conf "github.com/elastic/elastic-agent-libs/config"
@@ -303,4 +304,31 @@ func TestExtractArrayProcessor_Run(t *testing.T) {
 		assert.Equal(t, expMeta, result.Meta)
 		assert.Equal(t, event.Fields, result.Fields)
 	})
+}
+
+// TestExtractArrayFailOnErrorSafety verifies that when FailOnError=true and
+// the mapping index is out of bounds, the event fields are unchanged
+// (proving the clone skip is safe).
+func TestExtractArrayFailOnErrorSafety(t *testing.T) {
+	processor, err := New(conf.MustNewConfigFrom(mapstr.M{
+		"field": "array",
+		"mappings": mapstr.M{
+			"dest": 999,
+		},
+	}), logptest.NewTestingLogger(t, ""))
+	require.NoError(t, err)
+
+	input := mapstr.M{
+		"array": []interface{}{"only-one-element"},
+	}
+	event := &beat.Event{Fields: input.Clone()}
+	original := input.Clone()
+
+	result, err := processor.Run(event)
+	require.Error(t, err)
+	assert.Same(t, event, result)
+
+	result.Fields.Delete("error")
+	assert.Equal(t, original, result.Fields,
+		"event fields must be unchanged after out-of-bounds error (clone skip safety)")
 }
